backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-migrate==4.0.7
flask-smorest==0.46.1
flask-sqlalchemy==3.1.1
marshmallow==3.26.1
psycopg[binary]==3.2.6
PyJWT==2.10.1
requests==2.32.3
gunicorn==23.0.0
pytest==8.3.5
